pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "scourt-news-bot"
version = "0.1.0"
description = "Supreme Court press release collector and Teams notifier"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "apscheduler>=3.10.4,<4",
  "beautifulsoup4>=4.12.3,<5",
  "pdfplumber>=0.11.0,<1",
  "pypdf>=4.2.0,<6",
  "python-dotenv>=1.0.1,<2",
  "requests>=2.32.3,<3",
]

[project.scripts]
scourt-bot = "scourt_bot.main:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
